refactor(extension): change WalletRegisterStep folder name to WalletSetupNamePasswordStep
