+++ packages/core/src/ui/components/WalletSetup/WalletSetupNamePasswordStep/index.ts
