+++ packages/core/src/ui/components/WalletSetup/WalletSetupNamePasswordStep/types.ts
