+++ packages/core/src/ui/components/WalletSetup/WalletSetupNamePasswordStep/utils.ts
