+++ packages/core/src/ui/components/WalletSetup/index.ts
@@ -17,4 +17,4 @@ export * from './wallet-steps.common';
 export * from './wallet-steps';
 export * from './WalletSetupPasswordStep';
 export * from './MnemonicWordsWritedown';
-export * from './WalletRegisterStep';
+export * from './WalletSetupNamePasswordStep';
